# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/system/getServerInfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取服务器信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取服务器信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/getSystemConfig": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取配置文件内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取配置文件内容,返回包括系统配置",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SysConfigResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/reloadSystem": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "重启系统",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重启系统",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/setSystemConfig": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "设置配置文件内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header"
                    },
                    {
                        "description": "设置配置文件内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.System"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置配置文件内容",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.AliyunOSS": {
            "type": "object",
            "properties": {
                "accessKeyId": {
                    "type": "string"
                },
                "accessKeySecret": {
                    "type": "string"
                },
                "basePath": {
                    "type": "string"
                },
                "bucketName": {
                    "type": "string"
                },
                "bucketUrl": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                }
            }
        },
        "config.Autocode": {
            "type": "object",
            "properties": {
                "root": {
                    "type": "string"
                },
                "server": {
                    "type": "string"
                },
                "serverApi": {
                    "type": "string"
                },
                "serverInitialize": {
                    "type": "string"
                },
                "serverModel": {
                    "type": "string"
                },
                "serverRequest": {
                    "type": "string"
                },
                "serverRouter": {
                    "type": "string"
                },
                "serverService": {
                    "type": "string"
                },
                "transferRestart": {
                    "type": "boolean"
                },
                "web": {
                    "type": "string"
                },
                "webApi": {
                    "type": "string"
                },
                "webForm": {
                    "type": "string"
                },
                "webTable": {
                    "type": "string"
                }
            }
        },
        "config.AwsS3": {
            "type": "object",
            "properties": {
                "baseURL": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "pathPrefix": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretID": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
        "config.CORS": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                },
                "whitelist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.CORSWhitelist"
                    }
                }
            }
        },
        "config.CORSWhitelist": {
            "type": "object",
            "properties": {
                "allow-credentials": {
                    "type": "boolean"
                },
                "allow-headers": {
                    "type": "string"
                },
                "allow-methods": {
                    "type": "string"
                },
                "allow-origin": {
                    "type": "string"
                },
                "expose-headers": {
                    "type": "string"
                }
            }
        },
        "config.Captcha": {
            "type": "object",
            "properties": {
                "imgHeight": {
                    "description": "验证码高度",
                    "type": "integer"
                },
                "imgWidth": {
                    "description": "验证码宽度",
                    "type": "integer"
                },
                "is-return-captcha": {
                    "description": "是否返回验证码",
                    "type": "boolean"
                },
                "keyLong": {
                    "description": "验证码长度",
                    "type": "integer"
                }
            }
        },
        "config.Casbin": {
            "type": "object",
            "properties": {
                "modelPath": {
                    "description": "存放casbin模型的相对路径",
                    "type": "string"
                }
            }
        },
        "config.DB": {
            "type": "object",
            "properties": {
                "alias-name": {
                    "type": "string"
                },
                "config": {
                    "description": "高级配置",
                    "type": "string"
                },
                "dbname": {
                    "description": "数据库名",
                    "type": "string"
                },
                "disable": {
                    "type": "boolean"
                },
                "logMode": {
                    "description": "是否开启Gorm全局日志",
                    "type": "string"
                },
                "logZap": {
                    "type": "boolean"
                },
                "maxIdleConns": {
                    "description": "空闲中的最大连接数",
                    "type": "integer"
                },
                "maxOpenConns": {
                    "description": "打开到数据库的最大连接数",
                    "type": "integer"
                },
                "password": {
                    "description": "数据库密码",
                    "type": "string"
                },
                "path": {
                    "description": "服务器地址:端口",
                    "type": "string"
                },
                "port": {
                    "description": ":端口",
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "description": "数据库用户名",
                    "type": "string"
                }
            }
        },
        "config.Detail": {
            "type": "object",
            "properties": {
                "compareField": {
                    "description": "需要比较时间的字段",
                    "type": "string"
                },
                "interval": {
                    "description": "时间间隔",
                    "type": "string"
                },
                "tableName": {
                    "description": "需要清理的表名",
                    "type": "string"
                }
            }
        },
        "config.Email": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "收件人",
                    "type": "string"
                },
                "host": {
                    "description": "服务器地址",
                    "type": "string"
                },
                "isSSL": {
                    "description": "是否SSL",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "port": {
                    "description": "端口",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥",
                    "type": "string"
                },
                "to": {
                    "description": "收件人:多个以英文逗号分隔",
                    "type": "string"
                }
            }
        },
        "config.Excel": {
            "type": "object",
            "properties": {
                "dir": {
                    "type": "string"
                }
            }
        },
        "config.HuaWeiObs": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
        "config.JWT": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "description": "缓冲时间",
                    "type": "integer"
                },
                "expiresTime": {
                    "description": "过期时间",
                    "type": "integer"
                },
                "issuer": {
                    "description": "签发者",
                    "type": "string"
                },
                "signingKey": {
                    "description": "jwt签名",
                    "type": "string"
                }
            }
        },
        "config.Local": {
            "type": "object",
            "properties": {
                "path": {
                    "description": "本地文件路径",
                    "type": "string"
                }
            }
        },
        "config.Mysql": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "高级配置",
                    "type": "string"
                },
                "dbname": {
                    "description": "数据库名",
                    "type": "string"
                },
                "logMode": {
                    "description": "是否开启Gorm全局日志",
                    "type": "string"
                },
                "logZap": {
                    "description": "是否通过zap写入日志文件",
                    "type": "boolean"
                },
                "maxIdleConns": {
                    "description": "空闲中的最大连接数",
                    "type": "integer"
                },
                "maxOpenConns": {
                    "description": "打开到数据库的最大连接数",
                    "type": "integer"
                },
                "password": {
                    "description": "数据库密码",
                    "type": "string"
                },
                "path": {
                    "description": "服务器地址",
                    "type": "string"
                },
                "port": {
                    "description": "端口",
                    "type": "string"
                },
                "username": {
                    "description": "数据库用户名",
                    "type": "string"
                }
            }
        },
        "config.Pgsql": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "高级配置",
                    "type": "string"
                },
                "dbname": {
                    "description": "数据库名",
                    "type": "string"
                },
                "logMode": {
                    "description": "是否开启Gorm全局日志",
                    "type": "string"
                },
                "logZap": {
                    "description": "是否通过zap写入日志文件",
                    "type": "boolean"
                },
                "maxIdleConns": {
                    "description": "空闲中的最大连接数",
                    "type": "integer"
                },
                "maxOpenConns": {
                    "description": "打开到数据库的最大连接数",
                    "type": "integer"
                },
                "password": {
                    "description": "数据库密码",
                    "type": "string"
                },
                "path": {
                    "description": "服务器地址:端口",
                    "type": "string"
                },
                "port": {
                    "description": ":端口",
                    "type": "string"
                },
                "username": {
                    "description": "数据库用户名",
                    "type": "string"
                }
            }
        },
        "config.Qiniu": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "description": "秘钥AK",
                    "type": "string"
                },
                "bucket": {
                    "description": "空间名称",
                    "type": "string"
                },
                "imgPath": {
                    "description": "CDN加速域名",
                    "type": "string"
                },
                "secretKey": {
                    "description": "秘钥SK",
                    "type": "string"
                },
                "useCdnDomains": {
                    "description": "上传是否使用CDN上传加速",
                    "type": "boolean"
                },
                "useHttps": {
                    "description": "是否使用https",
                    "type": "boolean"
                },
                "zone": {
                    "description": "存储区域",
                    "type": "string"
                }
            }
        },
        "config.Redis": {
            "type": "object",
            "properties": {
                "addr": {
                    "description": "服务器地址:端口",
                    "type": "string"
                },
                "db": {
                    "description": "redis的哪个数据库",
                    "type": "integer"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "config.Server": {
            "type": "object",
            "properties": {
                "aliyunOSS": {
                    "$ref": "#/definitions/config.AliyunOSS"
                },
                "autoCode": {
                    "description": "auto",
                    "$ref": "#/definitions/config.Autocode"
                },
                "awsS3": {
                    "$ref": "#/definitions/config.AwsS3"
                },
                "captcha": {
                    "$ref": "#/definitions/config.Captcha"
                },
                "casbin": {
                    "$ref": "#/definitions/config.Casbin"
                },
                "cors": {
                    "description": "跨域配置",
                    "$ref": "#/definitions/config.CORS"
                },
                "db-list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.DB"
                    }
                },
                "email": {
                    "$ref": "#/definitions/config.Email"
                },
                "excel": {
                    "$ref": "#/definitions/config.Excel"
                },
                "huaWeiObs": {
                    "$ref": "#/definitions/config.HuaWeiObs"
                },
                "jwt": {
                    "$ref": "#/definitions/config.JWT"
                },
                "local": {
                    "description": "oss",
                    "$ref": "#/definitions/config.Local"
                },
                "mysql": {
                    "description": "gorm",
                    "$ref": "#/definitions/config.Mysql"
                },
                "pgsql": {
                    "$ref": "#/definitions/config.Pgsql"
                },
                "qiniu": {
                    "$ref": "#/definitions/config.Qiniu"
                },
                "redis": {
                    "$ref": "#/definitions/config.Redis"
                },
                "system": {
                    "$ref": "#/definitions/config.System"
                },
                "tencentCOS": {
                    "$ref": "#/definitions/config.TencentCOS"
                },
                "timer": {
                    "$ref": "#/definitions/config.Timer"
                },
                "zap": {
                    "$ref": "#/definitions/config.Zap"
                }
            }
        },
        "config.System": {
            "type": "object",
            "properties": {
                "addr": {
                    "description": "端口值",
                    "type": "integer"
                },
                "dbType": {
                    "description": "数据库类型:mysql(默认)|sqlite|sqlserver|postgresql",
                    "type": "string"
                },
                "env": {
                    "description": "环境值",
                    "type": "string"
                },
                "iplimitCount": {
                    "type": "integer"
                },
                "iplimitTime": {
                    "type": "integer"
                },
                "ossType": {
                    "description": "Oss类型",
                    "type": "string"
                },
                "useMultipoint": {
                    "description": "多点登录拦截",
                    "type": "boolean"
                }
            }
        },
        "config.TencentCOS": {
            "type": "object",
            "properties": {
                "baseURL": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "pathPrefix": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretID": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
        "config.Timer": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.Detail"
                    }
                },
                "spec": {
                    "description": "CRON表达式",
                    "type": "string"
                },
                "start": {
                    "description": "是否启用",
                    "type": "boolean"
                }
            }
        },
        "config.Zap": {
            "type": "object",
            "properties": {
                "director": {
                    "description": "日志文件夹",
                    "type": "string"
                },
                "encodeLevel": {
                    "description": "编码级",
                    "type": "string"
                },
                "format": {
                    "description": "输出",
                    "type": "string"
                },
                "level": {
                    "description": "级别",
                    "type": "string"
                },
                "logInConsole": {
                    "description": "输出控制台",
                    "type": "boolean"
                },
                "prefix": {
                    "description": "日志前缀",
                    "type": "string"
                },
                "showLine": {
                    "description": "显示行",
                    "type": "boolean"
                },
                "stacktraceKey": {
                    "description": "栈名",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.SysConfigResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/config.Server"
                }
            }
        },
...
        "system.System": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/config.Server"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}